+++ lib/src/reporters/code_climate/code_climate_reporter.dart
@@ -56,7 +56,7 @@ class CodeClimateReporter implements Reporter {
         result.add(CodeClimateIssue.maintainabilityIndex(
             func.firstLine,
             func.lastLine,
-            report.maintainabilityIndex,
+            report.maintainabilityIndex.toInt(),
             record.relativePath,
             key));
       }
